Added a widget to configure the workflow.
We use some of the settings already. We still do not use the token and phone id. We read these from environment variables. We also do not read tunnel information. But, feature is stable and complete enough for the main branch.

+++ Processes/qprocess_incoming_messages.py
@@ -9,7 +9,7 @@ import uvicorn
 
 from datetime import datetime
 
-from PyQt5.QtCore import QThread, pyqtSlot
+from PyQt5.QtCore import QThread, pyqtSlot as Slot
 from PyQt5.QtCore import pyqtSignal as Signal
 
 import redis.asyncio as redis
@@ -26,13 +26,13 @@ REDIS_PUBSUB_DB = os.getenv("PUBSUB_DB")
 class QIncomingMessagesProcessor(QThread):
     done = Signal()
 
-    def __init__(self, parent, port: int, run_date: datetime):
+    def __init__(self, parent, port: int):
         super().__init__(parent)
         self.port = port
         # Two tasks are schedule at this time.
         # Auto-decline opt-in messages. This is required because we need to physically print reports in the dead letter queue.
         # Collate the dead letter queue reports into so they are easier to print.
-        self.run_date = run_date
+        self.run_date = None
         self.server_proc = None
 
 
@@ -50,22 +50,32 @@ class QIncomingMessagesProcessor(QThread):
         await pim.process_messages(self.run_date)
 
 
-    async def sp_helper(self):
+    async def unsubscribe(self):
         r  = await redis.from_url("redis://localhost", db=REDIS_PUBSUB_DB)
         await r.publish(PENDING_DELIVERY_FILENAMES, "STOP")
         await r.publish(UPLOADED_ARTIFACTS, "STOP")
         await r.publish(OPT_IN_RESPONSES, "STOP")
 
-    @pyqtSlot()
+
+    @Slot()
     def shutdown_processes(self):
         logging.getLogger().info("...Shutting down the FastAPI server...")
-        asyncio.run(self.sp_helper())
+        asyncio.run(self.unsubscribe())
 
-        os.kill(self.server_proc.pid, signal.SIGTERM)
+        # poll() returns None if the process has not been terinated
+        if self.server_proc and self.server_proc.poll() is None:
+            os.kill(self.server_proc.pid, signal.SIGTERM)
 
-        self.server_proc.terminate()
-        self.server_proc.wait()
+            self.server_proc.terminate()
+            self.server_proc.wait()
         logging.getLogger().info("...FastAPI server was terminated...")
+
+
+    @Slot(datetime)
+    def start(self, run_date: datetime):
+        self.run_date = run_date
+
+        super().start()
 
 
     def run(self):
